# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinFamília API",
        "description": "REST API do FinFamília — gestão financeira familiar para o bot (smartphone).",
        "contact": {
            "email": "suporte@finfamilia.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/assets": {
            "get": {
                "tags": [
                    "Investments"
                ],
                "summary": "Lista ativos",
                "operationId": "5ecb46d908d0ea33df1904c896f35f67",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ativos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Investments"
                ],
                "summary": "Cria ativo",
                "operationId": "c3e5a860f52cecb9755abc964ca020a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "kind",
                                    "current_value"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "CDB"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "CDB Inter"
                                    },
                                    "kind": {
                                        "type": "string",
                                        "enum": [
                                            "renda_fixa",
                                            "fii",
                                            "acao",
                                            "etf",
                                            "previdencia"
                                        ]
                                    },
                                    "current_value": {
                                        "type": "string",
                                        "example": "10000"
                                    },
                                    "institution": {
                                        "type": "string"
                                    },
                                    "yield_percent": {
                                        "type": "string",
                                        "example": "120"
                                    },
                                    "yield_base": {
                                        "type": "string",
                                        "enum": [
                                            "cdi",
                                            "selic",
                                            "ipca",
                                            "prefixado"
                                        ]
                                    },
                                    "portfolio_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ativo criado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/assets/{asset}": {
            "delete": {
                "tags": [
                    "Investments"
                ],
                "summary": "Exclui ativo",
                "operationId": "01cef025b9f80e4a35de3ea35d4fb64f",
                "parameters": [
                    {
                        "name": "asset",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Excluído"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Investments"
                ],
                "summary": "Atualiza ativo",
                "operationId": "75a23a7fc8c985cd2a0e4624969229b3",
                "parameters": [
                    {
                        "name": "asset",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "current_value": {
                                        "type": "string"
                                    },
                                    "yield_percent": {
                                        "type": "string"
                                    },
                                    "portfolio_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ativo atualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/contributions": {
            "post": {
                "tags": [
                    "Investments"
                ],
                "summary": "Registra aporte/rendimento",
                "operationId": "a456a6e7fcd3d80bf1927b484afdc6bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kind",
                                    "amount",
                                    "occurred_on"
                                ],
                                "properties": {
                                    "kind": {
                                        "type": "string",
                                        "enum": [
                                            "aporte",
                                            "rendimento"
                                        ]
                                    },
                                    "amount": {
                                        "type": "string",
                                        "example": "1000"
                                    },
                                    "occurred_on": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "asset_id": {
                                        "type": "integer"
                                    },
                                    "account_id": {
                                        "type": "integer"
                                    },
                                    "portfolio_id": {
                                        "type": "integer"
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}